Finished swept collision from bullet to tank
Saved last X and Y position of bullet to determine where the bullet was going to end up. The collision check will calculate its next move based on its velocity then see if the bottom of the bullet is past the top two colliders of the tank and if its sides are within the two sides of the colliders of the tank.