Bullet Pool added. Can change # of bullets on the screen on the fly during simulation.
Bullet Pool added. User can change # of bullets to greater or less than the amount on screen currently.

Updated collision to create the illusion of the bullet hitting the player better at higher speeds.

Added # of bullets to slider in ImGUI.
Added ability to show wire frame for collision with checkbox in ImGUI.